Adds login controller
