+++ backend/controllers/auth.js
@@ -38,3 +38,29 @@ exports.signup = (req, res, next) => {
             res.status(500).json({ message: 'Creating the user failed.' });
         });
 };
+
+exports.login = (req, res, next) => {
+    const email = req.body.email;
+    const pw = req.body.password;
+    db.getDb().db().collection('users').findOne({email: email})
+        .then(userDoc => {
+            return bcrypt.compare(pw, userDoc.password);
+        })
+        .then(result => {
+            if (!result) {
+                throw Error();
+            };
+            const token = createToken();
+            res
+                .status(200)
+                .json({
+                    message: 'Authentication succeeded',
+                    token: token
+                });
+        })
+        .catch(err => {
+            res
+                .status(401)
+                .json({ message: 'Authentication failed, invalid username or password.' });
+        });
+};
